+++ main.go
@@ -4,31 +4,12 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"html/template"
 	_ "github.com/joho/godotenv/autoload"
 	"os"
 	"database/sql"
 	_ "github.com/lib/pq"
+	"./handlers"
 )
-
-type LoginPage struct {
-	Title string
-	User string
-	Pass string
-}
-
-
-func loginHandler(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		p := LoginPage{Title: "Login"}
-		t, _ := template.ParseFiles("templates/login.html")
-		t.Execute(w, p)
-	case "POST":
-		log.Println(r.FormValue("username"))
-		log.Println(r.FormValue("password"))
-	}
-}
 
 func main() {
 	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s " + "password=%s dbname=%s sslmode=disable", 
@@ -37,7 +18,7 @@ func main() {
 	defer db.Close()
 	port := "8080"
 	http.Handle("/", http.FileServer(http.Dir("static/")))
-	http.HandleFunc("/login", loginHandler)
+	http.HandleFunc("/login", handlers.LoginHandler)
 
 	log.Println("Server started on port " + port)
